Format struct initilizers in an all or nothing format.
Previously the UnwrappedLineParser always split struct initializers into serparate lines.
Now it does not, but instead the formatting of struct initializers are handled in
the continuation indenter. If the Struct initializer can fit on a single line,
it will place it there in a bin packing fashion. Otherwise a line break will be placed
between every field.

fixes #27